pyproject.toml
[project]
name = "sentiment-analysis-stocks"
version = "0.1.0"
description = ""
authors = [
    {name = "Tobias Sartori",email = "tobiasfaculdade@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.0,<7.0.0)",
    "toml-sort (>=0.24.2,<0.25.0)",
    "pymupdf (>=1.25.2,<2.0.0)",
    "pdfminer (>=20191125,<20191126)",
    "pdfplumber (>=0.11.5,<0.12.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "boto3 (>=1.36.21,<2.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


package-mode = false